added mapper

+++ src/main/java/org/dbrinkk/agenda/IAgendaRepository.java
@@ -2,6 +2,6 @@ package org.dbrinkk.agenda;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IAgendaRepository extends JpaRepository<Agenda, Integer> {
+public interface IAgendaRepository extends JpaRepository<Agenda, Long> {
 
 }
