+++ src/main/java/org/dbrinkk/agendaTodo/IAgendaTodoRepository.java
@@ -2,5 +2,5 @@ package org.dbrinkk.agendaTodo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IAgendaTodoRepository extends JpaRepository<AgendaTodo, Integer> {
+public interface IAgendaTodoRepository extends JpaRepository<AgendaTodo, Long> {
 }
