+++ src/main/java/org/dbrinkk/attachment/IAttachmentRepository.java
@@ -2,5 +2,5 @@ package org.dbrinkk.attachment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IAttachmentRepository extends JpaRepository<Attachment, Integer> {
+public interface IAttachmentRepository extends JpaRepository<Attachment, Long> {
 }
