+++ src/main/java/org/dbrinkk/todo/ITodoRepository.java
@@ -2,5 +2,5 @@ package org.dbrinkk.todo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ITodoRepository extends JpaRepository<Todo, Integer> {
+public interface ITodoRepository extends JpaRepository<Todo, Long> {
 }
