+++ src/main/java/org/dbrinkk/user/IUserRepository.java
@@ -2,5 +2,5 @@ package org.dbrinkk.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUserRepository extends JpaRepository<TodoUser, Integer> {
+public interface IUserRepository extends JpaRepository<TodoUser, Long> {
 }
